requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip==1.0
contourpy==1.2.0
cycler==0.12.1
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
flaskwebgui==1.1.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
gradio==3.35.2
gradio_client==0.13.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
linkify-it-py==2.0.3
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.9
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
orjson==3.9.15
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
scipy==1.12.0
seaborn==0.13.2
segment-anything==1.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
starlette==0.36.3
sympy==1.12
thop==0.1.1.post2209072238
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.3
ultralytics==8.0.120
urllib3==2.2.0
uvicorn==0.29.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.1
yarl==1.9.4
